Separate StringHelpers types into another files.
